refactor: 重構測試資料夾構造
issue: #40
