+++ tests/Feature/App/ExampleTest.php
@@ -1,14 +1,11 @@
 <?php
 
-namespace Tests\Api;
+namespace Tests\Feature\App;
 
-use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
 
 class ExampleTest extends TestCase
 {
-    use RefreshDatabase;
-
     /**
      * A basic test example.
      *
@@ -16,6 +13,8 @@ class ExampleTest extends TestCase
      */
     public function test_example()
     {
-        $this->assertTrue(true);
+        $response = $this->get('/');
+
+        $response->assertStatus(200);
     }
 }
